+++ spec/support/kitchen/terraform/config_attribute/root_module_directory_examples.rb
@@ -17,10 +17,10 @@
 require "support/kitchen/terraform/config_schemas/string_examples"
 
 ::RSpec
-  .shared_examples "Kitchen::Terraform::ConfigAttribute::Directory" do
+  .shared_examples "Kitchen::Terraform::ConfigAttribute::RootModuleDirectory" do
     include_context(
       "Kitchen::Terraform::ConfigSchemas::String",
-      attribute: :directory,
+      attribute: :root_module_directory,
       default_value: "kitchen_root"
     )
   end